Made the Alignments1 class *have* a GRangesList slot (@granges) instead of containing the GRangesList class.
This is cleaner and safer because Alignments1 objects then don't inherit the
GRangesList API out-of-the-box (most of the times, endomorphic methods in the
GRangesList API would return a broken Alignments1 object because they ignore
the cigar slot). It also avoids some conceptual conflicts inherent
to the GappedAlignments/GRangesList dual nature e.g. what start() should
return is ambiguous (integer vector or IntegerList object?).

Added the new granges() generic with methods for GappedAlignments objects.
It converts a GappedAlignments object into a GRangesList object. Note that, in
the case of Alignments1, this is just a slot extractor. ranges() is still
working and returning a CompressedNormalIRangesList object and the man page
now puts both -- granges() and ranges() -- side by side with granges() being
the preferred way to extract the ranges of a GappedAlignments object.


git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/Rsamtools@44830 bc3139a8-67e5-0310-9ffc-ced21a209358